main/livekit-server_simple/simple_requirements.txt
# Simplified requirements for basic LiveKit agent
livekit-agents>=1.2.0
livekit-plugins-groq>=0.1.0
livekit-plugins-elevenlabs>=0.1.0
livekit-plugins-deepgram>=0.1.0
livekit-plugins-cartesia>=0.1.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
edge-tts>=6.1.0
pyyaml>=6.0

# AWS SDK and Transcribe dependencies (for custom AWS STT provider)
boto3>=1.34.0
botocore>=1.34.0
amazon-transcribe>=0.6.0

# Custom VAD dependencies (replacing livekit-plugins-silero)
onnxruntime>=1.16.0
numpy>=1.24.0
pydantic>=2.0.0